requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file=requirements.txt --prerelease=allow --python=3.11 --emit-find-links
absl-py==2.1.0
    # via
    #   chex
    #   optax
    #   orbax-checkpoint
aiohappyeyeballs==2.5.0
    # via aiohttp
aiohttp==3.11.13
    # via gcsfs
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
attrs==25.1.0
    # via aiohttp
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.1
    # via requests
chex==0.1.89
    # via optax
click==8.1.8
    # via wandb
decorator==5.2.1
    # via gcsfs
docker-pycreds==0.4.0
    # via wandb
einops==0.8.1
    # via -r requirements.in
etils==1.12.0
    # via
    #   optax
    #   orbax-checkpoint
flax==0.10.4
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.3.0
    # via
    #   etils
    #   gcsfs
gcsfs==2025.3.0
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via wandb
google-api-core==2.24.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.38.0
    # via
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
google-auth-oauthlib==1.2.1
    # via gcsfs
google-cloud-core==2.4.2
    # via google-cloud-storage
google-cloud-storage==3.1.0
    # via gcsfs
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.69.1
    # via google-api-core
humanize==4.12.1
    # via orbax-checkpoint
idna==3.10
    # via
    #   requests
    #   yarl
importlib-resources==6.5.2
    # via etils
jax==0.5.2
    # via
    #   -r requirements.in
    #   chex
    #   flax
    #   optax
    #   orbax-checkpoint
jaxlib==0.5.1
    # via
    #   chex
    #   jax
    #   optax
jaxtyping==0.2.38
    # via -r requirements.in
libtpu==0.0.10.1
    # via jax
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.1
    # via
    #   jax
    #   jaxlib
    #   tensorstore
msgpack==1.1.0
    # via
    #   flax
    #   orbax-checkpoint
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.6.0
    # via orbax-checkpoint
numpy==2.2.3
    # via
    #   -r requirements.in
    #   chex
    #   flax
    #   jax
    #   jaxlib
    #   ml-dtypes
    #   optax
    #   orbax-checkpoint
    #   scipy
    #   tensorstore
    #   treescope
oauthlib==3.2.2
    # via requests-oauthlib
opt-einsum==3.4.0
    # via jax
optax==0.2.4
    # via flax
orbax-checkpoint==0.11.6
    # via
    #   -r requirements.in
    #   flax
pillow==11.1.0
    # via -r requirements.in
platformdirs==4.3.6
    # via wandb
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.0
    # via google-api-core
protobuf==5.29.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   orbax-checkpoint
    #   proto-plus
    #   wandb
psutil==7.0.0
    # via wandb
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.11.0a2
    # via wandb
pydantic-core==2.29.0
    # via pydantic
pygments==2.19.1
    # via rich
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   flax
    #   orbax-checkpoint
    #   wandb
requests==2.32.3
    # via
    #   gcsfs
    #   google-api-core
    #   google-cloud-storage
    #   jax
    #   requests-oauthlib
    #   wandb
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==13.9.4
    # via flax
rsa==4.9
    # via google-auth
scipy==1.15.2
    # via
    #   jax
    #   jaxlib
sentry-sdk==2.22.0
    # via wandb
setproctitle==1.3.5
    # via wandb
setuptools==75.8.2
    # via wandb
simplejson==3.20.1
    # via orbax-checkpoint
six==1.17.0
    # via docker-pycreds
smmap==5.0.2
    # via gitdb
tensorstore==0.1.72
    # via
    #   flax
    #   orbax-checkpoint
toolz==1.0.0
    # via chex
treescope==0.1.9
    # via flax
typing-extensions==4.12.2
    # via
    #   chex
    #   etils
    #   flax
    #   orbax-checkpoint
    #   pydantic
    #   pydantic-core
    #   wandb
urllib3==2.3.0
    # via
    #   requests
    #   sentry-sdk
wadler-lindig==0.1.3
    # via jaxtyping
wandb==0.19.8
    # via -r requirements.in
yarl==1.18.3
    # via aiohttp
zipp==3.21.0
    # via etils
